체중, 목표체중 소수점

+++ db/migrate/20201027045431_create_reports.rb
@@ -2,8 +2,8 @@ class CreateReports < ActiveRecord::Migration[6.0]
   def change
     create_table :reports do |t|
       t.string :name
-      t.integer :weight
-      t.integer :target_weight
+      t.float :weight
+      t.float :target_weight
       t.string :target_date
       t.float :morning_carbo
       t.float :morning_protein
